# src/openapi_server/models/account.py
# coding: utf-8

"""
    WASAPhoto Backend

    Keep in touch with your friends by sharing photos of special moments, thanks to WASAPhoto! You can upload your photos directly from your PC, and they will be visible to everyone following you.

    The version of the OpenAPI document: 1.0.0
    Contact: deodati.2025953@studenti.uniroma1.it
    Generated by OpenAPI Generator (https://openapi-generator.tech)

    Do not edit the class manually.
"""  # noqa: E501


from __future__ import annotations
import pprint
import re  # noqa: F401
import json




from pydantic import BaseModel, ConfigDict, Field
from typing import Any, ClassVar, Dict, List, Optional, Tuple, Union
from typing_extensions import Annotated
try:
    from typing import Self
except ImportError:
    from typing_extensions import Self

class Account(BaseModel):
    """
    An object representing a user account
    """ # noqa: E501
    id: Optional[Annotated[int, Field(strict=True, ge=0)]] = Field(default=None, description="ID represents an integer used as an identifier (such as postID and commentID)", alias="ID")
    pro_pic_b64: Optional[Union[Annotated[bytes, Field(strict=True, max_length=6990508)], Annotated[str, Field(strict=True, max_length=6990508)], Tuple[str, Annotated[bytes, Field(strict=True, max_length=6990508)]]]] = Field(default=None, description="An image encoded in base64", alias="proPicB64")
    followers: Optional[Annotated[int, Field(strict=True, ge=0)]] = Field(default=None, description="A counter used to represent likes/comments/followers counts")
    following: Optional[Annotated[int, Field(strict=True, ge=0)]] = Field(default=None, description="A counter used to represent likes/comments/followers counts")
    posts: Optional[List[Annotated[int, Field(strict=True, ge=0)]]] = Field(default=None, description="Posts by this user as postIDs")
    __properties: ClassVar[List[str]] = ["ID", "proPicB64", "followers", "following", "posts"]

    model_config = {
        "populate_by_name": True,
        "validate_assignment": True,
        "protected_namespaces": (),
    }


    def to_str(self) -> str:
        """Returns the string representation of the model using alias"""
        return pprint.pformat(self.model_dump(by_alias=True))

    def to_json(self) -> str:
        """Returns the JSON representation of the model using alias"""
        # TODO: pydantic v2: use .model_dump_json(by_alias=True, exclude_unset=True) instead
        return json.dumps(self.to_dict())

    @classmethod
    def from_json(cls, json_str: str) -> Self:
        """Create an instance of Account from a JSON string"""
        return cls.from_dict(json.loads(json_str))

    def to_dict(self) -> Dict[str, Any]:
        """Return the dictionary representation of the model using alias.

        This has the following differences from calling pydantic's
        `self.model_dump(by_alias=True)`:

        * `None` is only added to the output dict for nullable fields that
          were set at model initialization. Other fields with value `None`
          are ignored.
        """
        _dict = self.model_dump(
            by_alias=True,
            exclude={
            },
            exclude_none=True,
        )
        return _dict

    @classmethod
    def from_dict(cls, obj: Dict) -> Self:
        """Create an instance of Account from a dict"""
        if obj is None:
            return None

        if not isinstance(obj, dict):
            return cls.model_validate(obj)

        _obj = cls.model_validate({
            "ID": obj.get("ID"),
            "proPicB64": obj.get("proPicB64"),
            "followers": obj.get("followers"),
            "following": obj.get("following"),
            "posts": obj.get("posts")
        })
        return _obj