# src/openapi_server/models/post.py
# coding: utf-8

"""
    WASAPhoto Backend

    Keep in touch with your friends by sharing photos of special moments, thanks to WASAPhoto! You can upload your photos directly from your PC, and they will be visible to everyone following you.

    The version of the OpenAPI document: 1.0.0
    Contact: deodati.2025953@studenti.uniroma1.it
    Generated by OpenAPI Generator (https://openapi-generator.tech)

    Do not edit the class manually.
"""  # noqa: E501


from __future__ import annotations
import pprint
import re  # noqa: F401
import json




from datetime import datetime
from pydantic import BaseModel, ConfigDict, Field, field_validator
from typing import Any, ClassVar, Dict, List, Optional, Tuple, Union
from typing_extensions import Annotated
try:
    from typing import Self
except ImportError:
    from typing_extensions import Self

class Post(BaseModel):
    """
    An object representing a post
    """ # noqa: E501
    post_id: Optional[Annotated[int, Field(strict=True, ge=0)]] = Field(default=None, description="ID represents an integer used as an identifier (such as postID and commentID)", alias="postID")
    author: Optional[Annotated[int, Field(strict=True, ge=0)]] = Field(default=None, description="ID represents an integer used as an identifier (such as postID and commentID)")
    pub_time: Optional[datetime] = Field(default=None, description="Publication time for this post", alias="pubTime")
    image_b64: Optional[Union[Annotated[bytes, Field(strict=True, max_length=6990508)], Annotated[str, Field(strict=True, max_length=6990508)], Tuple[str, Annotated[bytes, Field(strict=True, max_length=6990508)]]]] = Field(default=None, description="An image encoded in base64", alias="imageB64")
    caption: Optional[Annotated[str, Field(min_length=0, strict=True, max_length=2048)]] = Field(default=None, description="A large text, used in post captions and comments")
    like_count: Optional[Annotated[int, Field(strict=True, ge=0)]] = Field(default=None, description="A counter used to represent likes/comments/followers counts", alias="likeCount")
    comments: Optional[List[Annotated[int, Field(strict=True, ge=0)]]] = Field(default=None, description="A list of comments (as commentIDs) belonging to a specific post")
    __properties: ClassVar[List[str]] = ["postID", "author", "pubTime", "imageB64", "caption", "likeCount", "comments"]

    @field_validator('caption')
    def caption_validate_regular_expression(cls, value):
        """Validates the regular expression"""
        if value is None:
            return value

        if not re.match(r"^.*?$", value):
            raise ValueError(r"must validate the regular expression /^.*?$/")
        return value

    model_config = {
        "populate_by_name": True,
        "validate_assignment": True,
        "protected_namespaces": (),
    }


    def to_str(self) -> str:
        """Returns the string representation of the model using alias"""
        return pprint.pformat(self.model_dump(by_alias=True))

    def to_json(self) -> str:
        """Returns the JSON representation of the model using alias"""
        # TODO: pydantic v2: use .model_dump_json(by_alias=True, exclude_unset=True) instead
        return json.dumps(self.to_dict())

    @classmethod
    def from_json(cls, json_str: str) -> Self:
        """Create an instance of Post from a JSON string"""
        return cls.from_dict(json.loads(json_str))

    def to_dict(self) -> Dict[str, Any]:
        """Return the dictionary representation of the model using alias.

        This has the following differences from calling pydantic's
        `self.model_dump(by_alias=True)`:

        * `None` is only added to the output dict for nullable fields that
          were set at model initialization. Other fields with value `None`
          are ignored.
        """
        _dict = self.model_dump(
            by_alias=True,
            exclude={
            },
            exclude_none=True,
        )
        return _dict

    @classmethod
    def from_dict(cls, obj: Dict) -> Self:
        """Create an instance of Post from a dict"""
        if obj is None:
            return None

        if not isinstance(obj, dict):
            return cls.model_validate(obj)

        _obj = cls.model_validate({
            "postID": obj.get("postID"),
            "author": obj.get("author"),
            "pubTime": obj.get("pubTime"),
            "imageB64": obj.get("imageB64"),
            "caption": obj.get("caption"),
            "likeCount": obj.get("likeCount"),
            "comments": obj.get("comments")
        })
        return _obj